+++ docs/examples/e12_slash_eh/main.py
@@ -35,6 +35,7 @@ async def slash_error(client, interaction_event, command, exception):
                                 SlasherApplicationCommandCategory,              |
                                 SlasherApplicationCommandFunction,              |
                                 SlasherApplicationCommandParameterAutoCompleter |
+                                FormSubmitCommand                                     |
         exception           :   BaseException                                   | The exception that was raised
 
     This should return a boolean